// src/gen/api/types/user_management_domain_dt_os_subscription_plan_dto.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * UserManagement.API
 * OpenAPI spec version: 1.0
 */
import type { ClicksiDataContractsCommonEnumsBusinessType } from "./clicksi_data_contracts_common_enums_business_type";
import type { ClicksiDataContractsCommonEnumsSubscriptionPermissions } from "./clicksi_data_contracts_common_enums_subscription_permissions";

export interface UserManagementDomainDTOsSubscriptionPlanDto {
  id: string;
  name: string;
  business_type: ClicksiDataContractsCommonEnumsBusinessType;
  monthly_price: number;
  yearly_price: number;
  description: string;
  permissions: ClicksiDataContractsCommonEnumsSubscriptionPermissions;
  max_connections: number;
  max_invites: number;
  max_products: number;
  api_calls_per_month: number;
  is_popular: boolean;
  trial_days: number;
  created_at: string;
}