// src/gen/api/types/user_management_domain_dt_os_team_subscription_dto.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * UserManagement.API
 * OpenAPI spec version: 1.0
 */
import type { UserManagementDomainDTOsSubscriptionPlanDtoNullable } from "./user_management_domain_dt_os_subscription_plan_dto_nullable";
import type { ClicksiDataContractsCommonEnumsSubscriptionStatus } from "./clicksi_data_contracts_common_enums_subscription_status";
import type { ClicksiDataContractsCommonEnumsBillingCycle } from "./clicksi_data_contracts_common_enums_billing_cycle";
import type { UserManagementDomainDTOsSubscriptionInvitationDto } from "./user_management_domain_dt_os_subscription_invitation_dto";

export interface UserManagementDomainDTOsTeamSubscriptionDto {
  id: string;
  /** @nullable */
  team_id?: string | null;
  plan_id: string;
  plan?: UserManagementDomainDTOsSubscriptionPlanDtoNullable;
  status: ClicksiDataContractsCommonEnumsSubscriptionStatus;
  billing_cycle: ClicksiDataContractsCommonEnumsBillingCycle;
  current_period_start: string;
  current_period_end: string;
  /** @nullable */
  trial_start?: string | null;
  /** @nullable */
  trial_end?: string | null;
  auto_renew: boolean;
  /** @nullable */
  cancelled_at?: string | null;
  /** @nullable */
  cancellation_effective_date?: string | null;
  /** @nullable */
  cancellation_reason?: string | null;
  /** @nullable */
  external_subscription_id?: string | null;
  created_at: string;
  invitations: UserManagementDomainDTOsSubscriptionInvitationDto[];
}